myfundoonote/requirements.txt
antiorm==1.2.1
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
db==0.1.1
db-sqlite3==0.0.1
Django==3.1.3
django-colorfield==0.3.2
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
isort==5.6.4
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.7
Pillow==8.0.1
psycopg2==2.8.6
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
python-decouple==3.3
pytz==2020.4
requests==2.25.0
ruamel.yaml==0.16.12
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
wrapt==1.12.1